add admin

+++ bookAPI/models.py
@@ -7,7 +7,7 @@ class Book(models.Model):
 
     title = models.CharField(_('title'), max_length=200)
     published_date = models.CharField(_('published date'), max_length=4)
-    average_rating = models.DecimalField(_('average rating'), default=0)
+    average_rating = models.FloatField(_('average rating'), default=0)
     ratings_count = models.IntegerField(_('rating count'), default=0)
     thumbnail = models.CharField(_('thumbnail'), max_length=300)
